+++ src/main/java/hudson/remoting/AbstractByteArrayCommandTransport.java
@@ -60,10 +60,8 @@ public abstract class AbstractByteArrayCommandTransport extends CommandTransport
         setup(new ByteArrayReceiver() {
             public void handle(byte[] payload) {
                 try {
-                    Command cmd = Command.readFrom(channel, new ObjectInputStreamEx(
-                            new ByteArrayInputStream(payload),channel.baseClassLoader,channel.classFilter));
+                    Command cmd = Command.readFrom(channel, payload);
                     receiver.handle(cmd);
-                    channel.notifyRead(cmd, payload.length);
                 } catch (IOException | ClassNotFoundException e) {
                     LOGGER.log(Level.WARNING, "Failed to construct Command in channel " + channel.getName(), e);
                 }